Move `revng/ABIAnalyses/Common.h` in EFA
