+++ include/revng/EarlyFunctionAnalysis/Common.h
