+++ lib/EarlyFunctionAnalysis/ABIAnalyses/MFIGraphs/ABIAnalysis.template
@@ -12,8 +12,8 @@
 #include "llvm/IR/Value.h"
 
 #include "revng/ABI/RegisterState.h"
-#include "revng/ABIAnalyses/Common.h"
 #include "revng/BasicAnalyses/GeneratedCodeBasicInfo.h"
+#include "revng/EarlyFunctionAnalysis/Common.h"
 #include "revng/EarlyFunctionAnalysis/EarlyFunctionAnalysis.h"
 #include "revng/MFP/MFP.h"
 #include "revng/Model/Binary.h"
